#160781 IllegalArgumentException: setSelectedIndex: 0 out of bounds

+++ cnd.makeproject/src/org/netbeans/modules/cnd/makeproject/ui/MakeLogicalViewProvider.java
@@ -77,6 +77,7 @@ import org.netbeans.modules.cnd.makeproject.api.actions.AddExistingItemAction;
 import org.netbeans.modules.cnd.makeproject.api.actions.NewFolderAction;
 import org.netbeans.modules.cnd.makeproject.api.configurations.BooleanConfiguration;
 import org.netbeans.modules.cnd.makeproject.api.configurations.Configuration;
+import org.netbeans.modules.cnd.makeproject.api.configurations.ConfigurationDescriptor.State;
 import org.netbeans.modules.cnd.makeproject.api.configurations.ConfigurationDescriptorProvider;
 import org.netbeans.modules.cnd.makeproject.api.configurations.Configurations;
 import org.netbeans.modules.cnd.makeproject.api.configurations.Folder;
@@ -153,14 +154,16 @@ public class MakeLogicalViewProvider implements LogicalViewProvider {
     }
 
     public Node createLogicalView() {
+        MakeConfigurationDescriptor configurationDescriptor = getMakeConfigurationDescriptor();
         if (ASYNC_ROOT_NODE) {
             log.fine("creating async root node in EDT? " + SwingUtilities.isEventDispatchThread());// NOI18N
-            return new MakeLogicalViewRootNode(getMakeConfigurationDescriptor().getLogicalFolders());
+            return new MakeLogicalViewRootNode(configurationDescriptor.getLogicalFolders());
         } else {
-            if (getMakeConfigurationDescriptor() == null) {
+            if (configurationDescriptor == null || configurationDescriptor.getState() == State.BROKEN
+                    || configurationDescriptor.getConfs().size() == 0) {
                 return new MakeLogicalViewRootNodeBroken();
             } else {
-                return new MakeLogicalViewRootNode(getMakeConfigurationDescriptor().getLogicalFolders());
+                return new MakeLogicalViewRootNode(configurationDescriptor.getLogicalFolders());
             }
         }
     }
@@ -582,18 +585,21 @@ public class MakeLogicalViewProvider implements LogicalViewProvider {
 
         @Override
         public Action[] getActions(boolean context) {
+            MakeConfigurationDescriptor descriptor = getMakeConfigurationDescriptor();
+
             // TODO: not clear if we need to call the following method at all
             // but we need to remove remembering the output to prevent memory leak;
             // I think it could be removed
-            if (gotMakeConfigurationDescriptor()) {
-                getMakeConfigurationDescriptor().getLogicalFolders();
+            if (descriptor != null) {
+                descriptor.getLogicalFolders();
             }
 
             List<Action> actions = new ArrayList<Action>();
             // Add standard actions
             Action[] standardActions;
-            MakeConfiguration active = (MakeConfiguration) getMakeConfigurationDescriptor().getConfs().getActive();
-            if (active.isMakefileConfiguration()) { // FIXUP: need better check
+            MakeConfiguration active = (descriptor == null)?
+                null : (MakeConfiguration) descriptor.getConfs().getActive();
+            if (descriptor == null || active == null || active.isMakefileConfiguration()) { // FIXUP: need better check
                 standardActions = getAdditionalDiskFolderActions();
             } else {
                 standardActions = getAdditionalLogicalFolderActions();
@@ -1139,9 +1145,10 @@ public class MakeLogicalViewProvider implements LogicalViewProvider {
                             copyItemConfigurations(movedItem.getItemConfigurations(), oldConfigurations);
                         }
                     } else {
-                        viewItemNode.getFolder().removeItem(item);
-                        toFolder.addItem(item);
-                        copyItemConfigurations(item.getItemConfigurations(), oldConfigurations);
+                        if (viewItemNode.getFolder().removeItem(item)) {
+                            toFolder.addItem(item);
+                            copyItemConfigurations(item.getItemConfigurations(), oldConfigurations);
+                        }
                     }
                 } else {
                     if (toFolder.isDiskFolder()) {
@@ -1151,8 +1158,9 @@ public class MakeLogicalViewProvider implements LogicalViewProvider {
                         String newName = IpeUtils.createUniqueFileName(toFolderPath, itemFO.getName(), itemFO.getExt());
                         FileObject movedFileFO = FileUtil.moveFile(itemFO, toFolderFO, newName);
                     } else if (IpeUtils.isPathAbsolute(item.getPath())) {
-                        viewItemNode.getFolder().removeItem(item);
-                        toFolder.addItem(item);
+                        if (viewItemNode.getFolder().removeItem(item)) {
+                            toFolder.addItem(item);
+                        }
                     } else if (item.getPath().startsWith("..")) { // NOI18N
                         String originalFilePath = FileUtil.toFile(viewItemNode.getFolder().getProject().getProjectDirectory()).getPath();
                         String newFilePath = FileUtil.toFile(toFolder.getProject().getProjectDirectory()).getPath();
@@ -1160,16 +1168,18 @@ public class MakeLogicalViewProvider implements LogicalViewProvider {
                         fromNewToOriginal = FilePathAdaptor.normalize(fromNewToOriginal);
                         String newPath = fromNewToOriginal + item.getPath();
                         newPath = IpeUtils.trimDotDot(newPath);
-                        viewItemNode.getFolder().removeItemAction(item);
-                        toFolder.addItemAction(new Item(FilePathAdaptor.normalize(newPath)));
+                        if (viewItemNode.getFolder().removeItemAction(item)) {
+                            toFolder.addItemAction(new Item(FilePathAdaptor.normalize(newPath)));
+                        }
                     } else {
                         Project toProject = toFolder.getProject();
                         FileObject fo = item.getFileObject();
                         FileObject copy = fo.copy(toProject.getProjectDirectory(), fo.getName(), fo.getExt());
                         String newPath = IpeUtils.toRelativePath(FileUtil.toFile(toProject.getProjectDirectory()).getPath(), FileUtil.toFile(copy).getPath());
-                        viewItemNode.getFolder().removeItemAction(item);
-                        fo.delete();
-                        toFolder.addItemAction(new Item(FilePathAdaptor.normalize(newPath)));
+                        if (viewItemNode.getFolder().removeItemAction(item)) {
+                            fo.delete();
+                            toFolder.addItemAction(new Item(FilePathAdaptor.normalize(newPath)));
+                        }
                     }
                 }
             } else if (type == DnDConstants.ACTION_COPY) {
